+++ lib/src/builder/focusable_field_builder.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_formy/src/builder/formy_builder.dart';
 
-import 'package:flutter_formy/src/models/controller/field_controller.dart';
+import 'package:flutter_formy/src/controller/field_controller.dart';
 import 'package:flutter_formy/src/models/field_state.dart';
 
 /// The type used by the [FocusableFieldBuilder]'s `builder` property.
@@ -59,7 +59,7 @@ typedef FocusableFieldWidgetBuilder<T> = Widget Function(
 /// ## See also
 ///
 /// * [FieldController], which manages the state of a field.
-/// * [FieldBuilder], a widget that rebuilds the interface based on changes to a [FieldController].
+/// * [FormyBuilder], the generic base class this widget extends.
 class FocusableFieldBuilder<T>
     extends FormyBuilder<FieldController<T>, FieldState<T>> {
   const FocusableFieldBuilder({