+++ lib/src/controller/field_config.dart
@@ -6,17 +6,15 @@ class FieldConfig<T> {
     this.validators = const [],
     this.initialValue,
     this.showErrorWhen = ShowError.whenIsTouched,
-    this.dependsOn = const [],
   });
   final String key;
   final List<FormyValidator<T>> validators;
   final T? initialValue;
   final ShowError showErrorWhen;
-  final List<String> dependsOn;
 
   FieldController<T> _initField(GroupController group) =>
       FieldController<T>._internal(
-          key, initialValue, showErrorWhen, validators, dependsOn, group);
+          key, initialValue, showErrorWhen, validators, group);
 }
 
 class FieldListConfig<T> extends FieldConfig<List<T>> {
@@ -25,11 +23,10 @@ class FieldListConfig<T> extends FieldConfig<List<T>> {
     super.initialValue,
     super.validators,
     super.showErrorWhen,
-    super.dependsOn,
   });
 
   @override
   FieldListControl<T> _initField(GroupController group) =>
       FieldListControl._internal(
-          key, initialValue, showErrorWhen, validators, dependsOn, group);
+          key, initialValue, showErrorWhen, validators, group);
 }