+++ lib/src/validators/formy_validator.dart
@@ -3,9 +3,8 @@ import 'package:flutter_formy/flutter_formy.dart';
 abstract class FormyValidator<T> {
   FormyValidator({required this.message});
   final String? message;
-  ValidationResult onValidate(FieldController<T> control);
+  ValidationResult onValidate(FieldController<T> controller);
 
-  ValidationResult call(FieldController<T> control) => onValidate(control);
-
-  final FormyFormManager formManager = FormyFormManager.instance;
+  ValidationResult call(FieldController<T> controller) =>
+      onValidate(controller);
 }