+++ lib/src/widgets/formy_dropdown.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_formy/src/builder/field_builder.dart';
-import 'package:flutter_formy/src/models/controller/field_controller.dart';
+import 'package:flutter_formy/src/controller/field_controller.dart';
 import 'package:flutter_formy/src/models/field_state.dart';
 import 'package:flutter_formy/src/models/validation_result.dart';