+++ lib/src/widgets/formy_option_mark_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_formy/src/builder/field_builder.dart';
-import 'package:flutter_formy/src/models/controller/field_controller.dart';
+import 'package:flutter_formy/src/controller/field_controller.dart';
 import 'package:flutter_formy/src/models/item_entry.dart';
 
 class FormyOptionListMarkWidget<ValueType, FieldType> extends StatelessWidget {