+++ test/src/builder/formy_builder_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_formy/src/manager/form_manager.dart';
 import 'package:flutter_formy/src/builder/formy_builder.dart';
-import 'package:flutter_formy/src/models/controller/field_controller.dart';
+import 'package:flutter_formy/src/controller/field_controller.dart';
 import 'package:flutter_formy/src/models/field_state.dart';
 import 'package:flutter_test/flutter_test.dart';